+++ project/mods/CertifiedPyro.InventoryTweaks/InventoryModded/inventory_constants.gd
@@ -1,0 +1,77 @@
+extends Reference
+
+const ITEM_ORDER := PoolStringArray([
+	# Item_Tools (gambling)
+	"scratch_off",
+	"scratch_off_2",
+	"scratch_off_3",
+	# Item_Tools (placeable fishing props)
+	"fish_trap",
+	"fish_trap_ocean",
+	"portable_bait",
+	# Item_Tools (catches from fishing)
+	"treasure_chest",
+	# Item_Consumables
+	"potion_catch",
+	"potion_catch_big",
+	"potion_catch_deluxe",
+	"potion_speed",
+	"potion_speed_burst",
+	"potion_beer",
+	"potion_wine",
+	"potion_revert",
+	"potion_small",
+	"potion_grow",
+	"potion_bounce",
+	"potion_bouncebig",
+	# Item_Tools (fishing rods)
+	"fishing_rod_simple",
+	"fishing_rod_skeleton",  # Spectral rod before collector's rods
+	"fishing_rod_travelers",
+	"fishing_rod_collectors",
+	"fishing_rod_collectors_shining",
+	"fishing_rod_collectors_glistening",
+	"fishing_rod_collectors_opulent",
+	"fishing_rod_collectors_radiant",
+	"fishing_rod_collectors_alpha",
+	"fishing_rod_prosperous",
+	# Item_Tools (dock shop)
+	"metal_detector",
+	# Item_Tools (misc in general store)
+	"boxing_glove",
+	"boxing_glove_super",
+	"hand_labeler",
+	"ringbox",
+	# Item_Chalk
+	"chalk_eraser",
+	"chalk_red",
+	"chalk_yellow",
+	"chalk_green",
+	"chalk_blue",
+	"chalk_white",
+	"chalk_black",
+	"chalk_special",
+	# Item_Tools (instruments)
+	"guitar",
+	"guitar_gradient",
+	"guitar_stickers",
+	"guitar_black",
+	"guitar_pink",
+	"guitar_gold",
+	"tambourine",
+	# Prop_Items
+	"prop_island_tiny",
+	"prop_island_med",
+	"prop_island_big",
+	# Creatures_MetalDetectLoot (pieces)
+	"mdl_piece_hat",
+	"mdl_piece_monocle",
+	"mdl_piece_sword",
+	"mdl_piece_watch",
+	# Creatures_MetalDetectLoot (junk)
+	"mdl_button",
+	"mdl_casing",
+	"mdl_coin",
+	"mdl_ring",
+	"mdl_sodatab"
+])